choose_disk: Removing cdrom & Read-only devices
If a device is read-only or a kind of cdrom, there is no need to keep it.
There is no chance at all such device could be used as Ceph device.

The read-only check could sounds like a duplicated case of the sr/cdrom
one but it's not.
- a cdrom device can be reported as RW
- a no sr/cdrom device can be read-only

So by having both checks, it does remove all those two categories of
devices.

Signed-off-by: Erwan Velu <erwan@redhat.com>